@@ -1,7 +1,10 @@
 #!/usr/bin/env node
-const fs = require("fs");
-const path = require("path");
+import fs from "fs";
+import path from "path";
+import { fileURLToPath } from "url";
 const filePath = process.argv[2];
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 if (!filePath) {
   console.error("No file path specified.");
