@@ -1,3 +1,3 @@
 pub mod app;
+pub mod audio_utils;
 pub mod extractor_music;
-// pub mod audio_utils;
